% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spike.R
\name{plot.lmls_spike}
\alias{plot.lmls_spike}
\title{Plot Method for lmls_spike object}
\usage{
\method{plot}{lmls_spike}(x, parameter = "location", plot_posterior = FALSE)
}
\arguments{
\item{x}{object of class spike_lmls}

\item{parameter}{takes a string which coefficients should be plotted, location or scale (default location)}

\item{plot_posterior}{Boolean whether to plot the posterior distributions or not (default FALSE). If False it plots inclusion probabilities.}
}
\description{
Plot Method for lmls_spike object
}
