% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_spike.R
\name{print.lmls_spike}
\alias{print.lmls_spike}
\title{Print method for class 'lmls_spike'}
\usage{
\method{print}{lmls_spike}(x, ...)
}
\arguments{
\item{x}{Object of class 'lmls_spike'}

<<<<<<< HEAD
\item{...}{Currently not in use}
=======
\item{...}{Currently not used.}
>>>>>>> c56ffa29256e53977acb064869dcddcc1c292490
}
\description{
Print method for class 'lmls_spike'
}
