% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike_slab.R
\name{spike_slab}
\alias{spike_slab}
\title{Title}
\usage{
spike_slab(
  m,
  a_theta_loc = 0.5,
  b_theta_loc = 0.5,
  a_tau_loc = 1,
  b_tau_loc = 1,
  a_theta_scl = a_theta_loc,
  b_theta_scl = b_theta_loc,
  a_tau_scl = a_tau_loc,
  b_tau_scl = b_tau_loc,
  v_0 = 0.01,
  nsim = 1000,
  burnin = round(nsim * 0.2),
  stepsize = sqrt(3) * (m$df)^(-1/6),
  prog_bar = TRUE,
  always_in_loc = NULL,
  always_in_scl = NULL,
  coef_init = NULL,
  seed = NULL
)
}
\arguments{
\item{m}{A location-scale regression model from the \code{\link[=lmls]{lmls()}} function.}

\item{a_theta_loc, b_theta_loc, a_tau_loc, b_tau_loc}{Hyperparameters for the location (see details).}

\item{a_theta_scl, b_theta_scl, a_tau_scl, b_tau_scl}{Hyperparameters for the scale (see details).}

\item{v_0}{Factor for the spike component.}

\item{nsim}{The number of MCMC samples to draw.}

\item{burnin}{Number of samples used as burn in phase, will be set to zero if negative.}

\item{stepsize}{The step size of the MMALA update.}

\item{prog_bar}{Show a progress bar?}

\item{always_in_loc, always_in_scl}{Select the variables which are NOT subject to selection.}

\item{coef_init}{Single value. If not NULL (default), the coefficients will be initialised with the given value.}

\item{seed}{If not NULL (default), the given value will be used for \link{set.seed}.}
}
\description{
Title
}
