% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike_slab_block.R
\name{spike_slab_block}
\alias{spike_slab_block}
\title{Title}
\usage{
spike_slab_block(
  m,
  a_theta_loc = 0.5,
  b_theta_loc = 0.5,
  a_tau_loc = 1,
  b_tau_loc = 1,
  a_theta_scl = a_theta_loc,
  b_theta_scl = b_theta_loc,
  a_tau_scl = a_tau_loc,
  b_tau_scl = b_tau_loc,
  v_0 = 0.01,
  nsim = 1000,
  burnin = round(nsim * 0.2),
  stepsize = sqrt(3) * (m$df)^(-1/6)
)
}
\arguments{
\item{m}{A location-scale regression model from the \code{\link[=lmls]{lmls()}} function.}

\item{a_theta_loc, b_theta_loc, a_tau_loc, b_tau_loc}{Hyperparameters for the location (see details).}

\item{a_theta_scl, b_theta_scl, a_tau_scl, b_tau_scl}{Hyperparameters for the scale (see details).}

\item{v_0}{Factor for the spike component.}

\item{nsim}{The number of MCMC samples to draw.}

\item{burnin}{Number of samples used as burn in phase, will be set to zero if negative.}

\item{stepsize}{The step size of the MMALA update.}
}
\description{
Title
}
