% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.lmls_spike}
\alias{summary.lmls_spike}
\title{Summary for spike and slab priors}
\usage{
\method{summary}{lmls_spike}(object, ...)
}
\arguments{
\item{object}{Object of class lmls_spike}

<<<<<<< HEAD
\item{...}{Currently not in use}
=======
\item{...}{Currently not used.}
>>>>>>> c56ffa29256e53977acb064869dcddcc1c292490
}
\description{
Summary for spike and slab priors
}
