% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spike.R
\name{plot.lmls_spike}
\alias{plot.lmls_spike}
\title{Plot Method for lmls_spike object}
\usage{
\method{plot}{lmls_spike}(x, parameter = "location", plot_type = "inclusion", ...)
}
\arguments{
\item{x}{Object of class spike_lmls}

\item{parameter}{Takes a string which coefficients should be plotted, 'location' (default) or 'scale', can be abbreviated.}

\item{plot_type}{Can be one of 'inclusion' (default), 'posterior' or 'random walk', can be abbreviated.}

<<<<<<< HEAD
\item{...}{Currently not in use}
=======
\item{...}{Currently not used.}
>>>>>>> c56ffa29256e53977acb064869dcddcc1c292490
}
\description{
Plot Method for lmls_spike object
}
