% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_spike.R
\name{print.lmls_spike}
\alias{print.lmls_spike}
\title{Print method for class 'lmls_spike'}
\usage{
\method{print}{lmls_spike}(x, ...)
}
\arguments{
\item{x}{Object of class 'lmls_spike'}
}
\description{
Print method for class 'lmls_spike'
}
