% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.lmls_spike}
\alias{summary.lmls_spike}
\title{Summary for spike and slab priors}
\usage{
\method{summary}{lmls_spike}(object)
}
\arguments{
\item{object}{Object of class lmls_spike}

\item{...}{Currently not used}
}
\description{
Summary for spike and slab priors
}
